requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
mysql-connector-python
python-dotenv
joblib